TestOrders

+++ Droneshop.Core/ApplicationService/Services/OrderService.cs
@@ -17,11 +17,17 @@ namespace Droneshop.Core.ApplicationService.Services
         }
         public Order CreateOrder(Order order)
         {
+            ValidateData(order);
             return _orderRepo.CreateOrder(order);
         }
 
         public Order DeleteOrder(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException("The entered id has to be atleast 1");
+            }
+            
             return _orderRepo.DeleteOrder(id);
         }
 
@@ -32,12 +38,50 @@ namespace Droneshop.Core.ApplicationService.Services
 
         public Order ReadOrderById(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException("The entered id has to be atleast 1");
+            }
+
+            var orderFound = _orderRepo.ReadOrderById(id);
+
+            if (orderFound == null)
+            {
+                throw new ArgumentException("Couldn't find any orders with the entered id");
+            }
+            
             return _orderRepo.ReadOrderById(id);
         }
 
         public Order UpdateOrder(Order order)
         {
+            ValidateData(order);
             return _orderRepo.UpdateOrder(order);
+        }
+
+        private void ValidateData(Order order)
+        {
+            if (order.Id < 1)
+            {
+                throw new ArgumentException("The entered id has to be atleast 1");
+            }
+
+            if (order.OrderDate == null)
+            {
+                throw new ArgumentException("OrderDate Cannot Be Null or Empty");
+            }
+
+            if (order.Customer == null)
+            {
+                throw new ArgumentException("Customer Cannot Be Null or Empty");
+            }
+
+            if (order.OrderLines == null)
+            {
+                throw new ArgumentException("OrderLine Cannot Be Null or Empty");
+            }
+
+           
         }
     }
 }
